requirements.txt
aiometer>=0.3.0
fastapi>=0.79.0
httpx>=0.23.0
loguru>=0.6.0
pyhumps>=3.7.2
python-dotenv>=0.20.0
uvicorn>=0.18.2